chore: tweak some config, expose "getStaticFiles"
