+++ packages/remix-fastify/src/index.ts
@@ -3,3 +3,4 @@ import "./globals";
 export type { GetLoadContextFunction, RequestHandler } from "./server";
 export { createRequestHandler } from "./server";
 export { remixFastifyPlugin } from "./plugin";
+export { getStaticFiles } from "./utils";
